Remove a problem test from 'inherit' test case
There is a problem with reading an anonymous record type, resulting in an error
such as "ERROR:  input of anonymous composite types is not implemented".

This is an existing issue affecting even XL 9.5, but was masked so far by lack
of testing and also because we inadvertently accepted wrong output in XL 9.5.
For now, move the case to xl_known_bugs and open an issue to track this
potential long standing bug.