Amend and accept regression diffs in 'insert' test case
We don't support triggers and hence the do-nothing trigger doesn't have the
desired effect in XL. As a result, the inserts are not skipped in XL (while
they are skipped because the before-row trigger returns nothing).  Also we
expect columns to be in the same order for partitions. So make those changes
too.