Provide enough space for knownXids array
Commit be7777d2a9d8d70c931e80868f4f7e373f3f98f8 rearranged things so that the
intial XID value is set appropriately in the GTM control file. But that seems
to have caused problem with a standby setup and the array to hold knownXids is
no longer enough. Hence make some adjustments there.

This was stopping a standby from starting up correctly as observed during the
tests.