Initialise PGXCNodeIdentifier as part of multi-executor setup
We used to do this while creating a portal. It's not clear why that place was
chosen by the original author, but it seems more logical to do so while setting
up multi-node execution environment. This also solves the problem of a parallel
worker not seeing the correct state, as reported by Virendra Kumar.