Overhaul CREATE BARRIER handling
Instead of using a low-level LWLock, the new code uses a high-level advisory
lock (though we some tweaks). This ensures that the command doesn't have to
hold a low-level lock across network roundtrips as the old implementation
required. This also ensures that the command responds to various timeouts such
as statement_timeout and lock_timeout instead of remaining blocked forever.

While reworking, I also realised that the current implementation wasn't
actually doing the right thing to stop in-flight 2PC transactions while a
BARRIER is being generated. This is because we were merely acquiring the lock
when a COMMIT TRANSACTION command is sent to the remote nodes. But that doesn't
participate in a 2PC. So now correctly acquire the lock before 2PC is started.
The lock is released with the transaction and hence no special handling is
required to release the lock correctly. (This also takes care of any errors
and/or timeouts while holding the lock.)

We should really add some test coverage for BARRIER, but this commit does not
do that.