Do not miss opportunity to FQS certain joins
When a query involves a 3-way (or N-way) join between two distributed tables and
a replicated table, such joins should be shippable to datanodes as long as the
join between the distributed tables is an equi-join on the distributed column,
with the same distribution property AND the join between the result of the join
on distributed tables and the replicate table is either an INNER JOIN or a LEFT
JOIN where the right side is the replicated table.

This fix allows such joins to be pushed down to the datanodes. A few test cases
have been added to test this functionality.