Use a separate buffer to copy GTM options before starting it
We are seeing an inexplicable failure while testing RPM packages. The GTM proxy
fails to start as it tries to access the configuration file in the current
working directory instead of the GTM proxy's data directory. Some debug tracing
indicates that the problem could be with sprintf trying to copy the source
buffer into destination and breaking along the way. So avoid doing that and see
if the packages like this.