Use timeouts in communications between GTM proxy and GTM.
We now have a new comm_timeout connection parameter while connecting to the
GTM/GTM Proxy. The default value for the parameter is 60s. If the client
doesn't get a response within the timeout, it errors out. Most GTM clients at
this point use compile time default, but we can later turn this into
configurable parameter.

This may help resolve certain conditions where a proxy infinitely waits for a
response from the GTM, because of some protocol error, thus freezing the
entire node.