Make comm_timeout configurable via GTM proxy GUC
The default is now set to 0, which means infinite wait and matches the
behaviour before the comm_timeout was introduced via
f04b2efc3dbf93b0d004b5ea0493fcfa0046a35f.

This patch introduces 2 GUCs on the datanode/coordinator side:

gtm_connect_timeout - timeout in seconds for GTM/GTM proxy connection
establishment. The default value of this GUC is 60s

gtm_comm_timeout - timeout in seconds to wait for a response from the
GTM/GTM Proxy. The default value of thie GUC is 0, which means wait
forever

This patch also introduces 1 GUC on the GTM proxy side:

comm_timeout - timeout in seconds to wait for a response from the GTM. The
default value of this GUC is 0, which means wait forever.

So while the default behaviour remains the same as before the patch, this
allows user to tune the values to avoid unpleasant cluster behaviour.