Do not report txn commit/abort to the GTM via parallel worker.
We've seen a race condition where a parallel worker process on the coordinator
pre-maturely reports txn status to the GTM, resulting in cleanup of the txn
from the GTM. This is wrong and we must let the leader process do this.

This may not fully explain some of the recent reports we have seen with
catalog corruptions where an updated row is marked visible as its xmax is seen
as aborted. But this looks like a sane fix anyways. So push it while we
continue to investigate the reported problem.