Add an infrastructure to store and dump debug trace messages.
These messages are kept in a thread-specific info and managed in a circular
queue. So only last N messages are ever tracked. They can be dumped to an
external file by invoking corresponding dump function. Right now, there is no
API to invoke the function and we may need to do so via a debugger.

A new GUC parameter debug_buffers is added which can be set in GTM and GTM
Proxy's configuration file.

The buffers can be dumped via dumpGTMDebugBuffers() and
dumpGTMProxyDebugBuffers() functions. These functions take a single boolean
parameter. When set to "true", only current thread's buffers are dumped. When
set to "false", all thread's buffers are dumped.

Currently the output is dumped to '/tmp/gtm_debug.log' and
'/tmp/gtm_proxy_debug.log' files and there is no way to change that. This might
change in future.

The logging API addGTMDebugMessage() looks very similar to elog() API. This can
be used to log to the buffers as well as elog at the desired level. Messages
are added to the buffers irrespective of the specified log level, but added to
the log only if log_min_messages accepts the given log level.

With the new logging infrastructure, some tracing messages are added with this
patch, but more are expected. We also exchange some additional information
between GTM proxy and the GTM and log those for tracing purposes. Arguably,
this increases communication overhead and should ideally be made optional,
either via config or at compile time.