Revert "Fetch minmxid from remote nodes and compute local value"
This reverts commit d664bec89cb316ea7e1ece546d3402b69596ba6c.

It wasn't correct to fetch multi XID values from remote nodes and then compute
coordinator side relmminxid as the minimum of the remote values. By doing so,
the pg_class.relmminxid may become greater than the multiXID value known to the
local instance. This then causes problems because the code to update database's
frozenXID considers that as an ERROR condition. Also, since multiXID values are
not maintained by the GTM and are not cluster-wide, it makes no sense to fetch
remote values and use them to compute local state.

Per report from Hengbing.